Behrouz Biryani front page clone initial stuff
This commit contains the initial version of the Behrouz Biryani homepage clone. Built using only HTML and CSS, the goal of this project is to replicate the visual structure and styling of the original website’s front page.

It includes a basic layout with images, fonts, and sections arranged to match the real UI as closely as possible (without interactivity or backend). This project is intended for practice and improving frontend development skills.